requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
idna==3.10
pycparser==2.22
PyGithub==2.7.0
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.5
typing_extensions==4.14.1
urllib3==2.5.0
